requirements.txt
# Collecte de données
requests==2.31.0

# Base de données
pymongo==4.6.1
psycopg2-binary==2.9.11
sqlalchemy==2.0.25

# API Backend
pydantic>=2.5.0,<2.6.0
fastapi>=0.109.0,<0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Tests
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilitaires
python-dotenv==1.0.0
python-dateutil==2.8.2
